requirements.txt
pdfminer.six
python-docx
pytesseract
Pillow
nltk
transformers
streamlit
